+++ fleetpings/admin.py
@@ -35,6 +35,7 @@ class FleetDoctrineAdmin(admin.ModelAdmin):
 
     list_display = ("name", "link", "notes", "is_enabled")
     list_filter = ("is_enabled",)
+    filter_horizontal = ("restricted_to_group",)
     ordering = ("name",)
 
 
@@ -70,6 +71,7 @@ class FleetTypeAdmin(admin.ModelAdmin):
 
     list_display = ("name", "embed_color", "notes", "is_enabled")
     list_filter = ("is_enabled",)
+    filter_horizontal = ("restricted_to_group",)
     ordering = ("name",)