+++ fleetpings/views.py
@@ -64,10 +64,26 @@ def index(request):
         )
 
     # get fleet types
-    fleet_types = FleetType.objects.filter(is_enabled=True).order_by("name")
+    fleet_types = (
+        FleetType.objects.filter(
+            Q(restricted_to_group__in=request.user.groups.all())
+            | Q(restricted_to_group__isnull=True),
+            is_enabled=True,
+        )
+        .distinct()
+        .order_by("name")
+    )
 
     # get doctrines
-    doctrines = FleetDoctrine.objects.filter(is_enabled=True).order_by("name")
+    doctrines = (
+        FleetDoctrine.objects.filter(
+            Q(restricted_to_group__in=request.user.groups.all())
+            | Q(restricted_to_group__isnull=True),
+            is_enabled=True,
+        )
+        .distinct()
+        .order_by("name")
+    )
 
     # get formup locations
     formup_locations = FormupLocation.objects.filter(is_enabled=True).order_by("name")